+++ gatsby-site/src/components/Page/Page.css
@@ -5,6 +5,16 @@
     --background: rgba(255,255,255,0);
     --leftmargin: 80px;
 }
+
+.page-container {
+    margin-top: 60px;
+    text-align: center;
+    width: 100%;
+}
+
+.page {
+    margin-top: 60px;
+}
    
 .page_title {
     font-family: var(--colarkfont);
@@ -12,10 +22,8 @@
 }
 
 img.logo {
-    font-family: var(--colarkfont);
-    color: var(--colarkfontcolor);
     height: 100%;
-    width: 20%;
+    width: 300px;
     border-radius: 10px;
     text-align: center;
     margin: 0 auto;
@@ -31,3 +39,13 @@ img.logo {
     font-family: var(--colarkfont);
     color: var(--colarkfontcolor);
 }
+
+@media only screen and (max-width: 535px) {
+    .page-container {
+        width: 80% !important;
+    }
+    .img.logo {
+      margin: 0 auto;
+      text-align: center;
+    }
+  }
