--- gatsby-site/src/components/Phase0/Phase0.css
@@ -1,8 +1,0 @@
-.page-container {
-    margin-top: 60px;
-    text-align: center;
-}
-
-.page {
-    margin-top: 60px;
-}
